Code
These python and matlab scripts are used to do things like process data, convert video to pngs, and build the training/testing/verification database.
The swift files are used to run the neural network on an iPhone. The coreML.mlmodel is not included here because of the upload size limit on GitHub.